Update proj3_phase2_draft10.py
added ros node to publish commands to robot in gazebo. refactored code to make it modular. Added the option to use weighted A*.  Adjusted animation to not stop program when user closes window but to start ros port after it.  reran A* after the animation with the user input goal_x and goal_y, and fixed the other parameters.